requirements.txt
streamlit
streamlit-authenticator
mysql-connector-python
python-dotenv
groq
langchain-text-splitters
sentence-transformers
faiss-cpu
python-dateutil
bcrypt